+++ src/backend/core/migrations/0013_active_unnaccent_postgres_extension.py
@@ -1,0 +1,14 @@
+from django.db import migrations
+
+from django.contrib.postgres.operations import UnaccentExtension
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('core', '0012_item_malware_detection_info'),
+    ]
+
+    operations = [
+        UnaccentExtension(),
+    ]